Paths + Utilities + Config
- Initial support for paths
- Initial support for utilities
- Configuration stored to home path

+++ Modules/Powerd/Powerd-Development.psm1
@@ -1,16 +1,34 @@
 
 #Initialize Developer Environment from VS 2015 batch file
-function Initialize-DeveloperEnvironment {
-	$tempFile = [IO.Path]::GetTempFileName();
-	$batchFile = '"C:\Program Files (x86)\Microsoft Visual Studio 14.0\Common7\Tools\VsDevCmd.bat"';
-	$tempFileName = '"' + $tempFile + '"';
-	cmd /c " $batchFile && set > $tempFileName";
-	Get-Content $tempFile | ForEach-Object {
-		if($_ -match "^(.*?)=(.*)$"){
-			Set-Content "env:\$($matches[1])" $matches[2];
-		}
+function Initialize-Development {
+	$batchFile = [IO.FileInfo]"C:\Program Files (x86)\Microsoft Visual Studio 14.0\Common7\Tools\VsDevCmd.bat";
+	if(! $batchFile.Exists) {
+		return;
 	}
-	Remove-Item $tempFile;
+
+	$tempFile = [IO.Path]::GetTempFileName();
+	cmd /c " `"$batchFile`" && set > `"$tempFile`"";
+	try {
+		Get-Content $tempFile | ForEach-Object {
+			if($_ -match "^(.*?)=(.*)$"){
+				Set-Content "env:\$($matches[1])" $matches[2];
+			}
+		}	
+	}
+	finally {
+		Remove-Item $tempFile;		
+	}
 }
 
-Initialize-DeveloperEnvironment;
+function Write-GitPrompt {
+    $realLASTEXITCODE = $LASTEXITCODE;
+
+    # Reset color, which can be messed up by Enable-GitColors
+    $Host.UI.RawUI.ForegroundColor = $GitPromptSettings.DefaultForegroundColor;
+    Write-Host($pwd.ProviderPath) -NoNewline;
+    Write-VcsStatus;
+    $global:LASTEXITCODE = $realLASTEXITCODE;
+    return "> ";
+}
+
+Initialize-Development;
